requirements.txt
astunparse==1.6.3
black==25.1.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.7.0
google-pasta==0.2.0
grpcio==1.74.0
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
mypy_extensions==1.1.0
namex==0.1.0
networkx==3.4.2
numpy==2.1.3
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pathspec==0.12.1
pika==1.3.2
pillow==11.3.0
platformdirs==4.4.0
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
Pygments==2.19.2
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.1.0
scipy==1.15.3
six==1.17.0
sng4onnx==1.0.4
supervision==0.26.1
sympy==1.14.0
termcolor==3.1.0
tomli==2.2.1
torch
torchvision
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.176
ultralytics-thop==2.0.15
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
